requirements.txt
fastapi
uvicorn[standard]
mangum
sqlalchemy
aiomysql
pydantic
python-dotenv
pytest
pytest-asyncio
httpx
pytest-mock
